requirements.txt
discord.py
pytesseract
opencv-python
numpy
Pillow
